Extract the delegation audit test helpers
The two delegation audit tests were ~90-line near-clones differing only in
the token minted and the final assertion, so every new case would have
been another copy. Move them to their own file and factor the shared
setup into signTestJWT, exchangeForDelegatedToken and auditEventForToken.

Mechanical: the assertion sets are unchanged, including failure messages,
which the helpers carry through as a parameter. The only addition is
asserting the auditor's Close error, and the switch from defer to
t.Cleanup that the parallel-test rule asks for.
