+++ pkg/authserver/integration_test.go
@@ -15,8 +15,6 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"net/url"
-	"os"
-	"path/filepath"
 	"strings"
 	"sync/atomic"
 	"testing"
@@ -31,7 +29,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/stacklok/toolhive/pkg/audit"
 	"github.com/stacklok/toolhive/pkg/auth"
 	"github.com/stacklok/toolhive/pkg/auth/upstreamtoken"
 	servercrypto "github.com/stacklok/toolhive/pkg/authserver/server/crypto"
@@ -790,202 +787,6 @@ func TestIntegration_TokenExchange_ConfidentialClientHappyPath(t *testing.T) {
 	require.True(t, ok, "exp claim should be a number")
 	assert.WithinDuration(t, now.Add(15*time.Minute), time.Unix(int64(exp), 0), 2*time.Minute,
 		"delegated token exp must be capped at the 15m delegation lifespan, not the subject token's 30m")
-}
-
-// readSingleAuditEvent reads the audit log file at path and decodes exactly
-// one newline-delimited JSON event, failing the test if there isn't exactly one.
-func readSingleAuditEvent(t *testing.T, path string) map[string]any {
-	t.Helper()
-
-	data, err := os.ReadFile(path)
-	require.NoError(t, err)
-	require.NotEmpty(t, strings.TrimSpace(string(data)), "audit log is empty; expected one event")
-
-	lines := strings.Split(strings.TrimSpace(string(data)), "\n")
-	require.Len(t, lines, 1, "expected exactly one audit event, got: %q", string(data))
-
-	var event map[string]any
-	require.NoError(t, json.Unmarshal([]byte(lines[0]), &event))
-	return event
-}
-
-// TestIntegration_TokenExchange_AuditDelegationChain proves the end-to-end
-// audit story for RFC 8693 delegation: a delegated token minted by the real
-// token-exchange handler is validated by the auth middleware, and the audit
-// middleware emits an event whose delegationChain records the acting agent.
-func TestIntegration_TokenExchange_AuditDelegationChain(t *testing.T) {
-	t.Parallel()
-
-	const (
-		agentClientID     = "test-audit-agent-client"
-		agentClientSecret = "test-audit-agent-secret"
-		delegatedUserSub  = "audit-delegated-user-sub"
-	)
-
-	agentClient, err := registration.New(registration.Config{
-		ID:         agentClientID,
-		Secret:     agentClientSecret,
-		Public:     false,
-		GrantTypes: []string{oauthproto.GrantTypeTokenExchange},
-		Scopes:     registration.DefaultScopes,
-		Audience:   []string{testAudience},
-	})
-	require.NoError(t, err)
-
-	m := startMockOIDC(t)
-	ts := setupTestServerWithMockOIDC(t, m, withExtraClient(agentClient))
-
-	// Mint the subject token (user) signed by the server's key.
-	signer, err := jose.NewSigner(
-		jose.SigningKey{Algorithm: jose.RS256, Key: ts.PrivateKey},
-		(&jose.SignerOptions{}).WithType("JWT").WithHeader("kid", "test-key"),
-	)
-	require.NoError(t, err)
-
-	now := time.Now()
-	subjectToken, err := jwt.Signed(signer).
-		Claims(jwt.Claims{
-			Issuer:   testIssuer,
-			Subject:  delegatedUserSub,
-			Audience: jwt.Audience{testAudience},
-			Expiry:   jwt.NewNumericDate(now.Add(30 * time.Minute)),
-			IssuedAt: jwt.NewNumericDate(now),
-		}).
-		Claims(map[string]any{
-			"client_id": agentClientID,
-		}).
-		Serialize()
-	require.NoError(t, err)
-
-	// Exchange it for a delegated token carrying act.sub = agent.
-	resp := makeTokenRequest(t, ts.Server.URL, url.Values{
-		"grant_type":         {oauthproto.GrantTypeTokenExchange},
-		"subject_token":      {subjectToken},
-		"subject_token_type": {oauthproto.TokenTypeAccessToken},
-		"client_id":          {agentClientID},
-		"client_secret":      {agentClientSecret},
-	})
-	defer resp.Body.Close()
-	body := parseTokenResponse(t, resp)
-	require.Equal(t, http.StatusOK, resp.StatusCode,
-		"token exchange should succeed, got %d (body: %v)", resp.StatusCode, body)
-	delegated, ok := body["access_token"].(string)
-	require.True(t, ok, "access_token should be a string")
-	require.NotEmpty(t, delegated)
-
-	// Build a validator that trusts the server's issuer and keys, exactly like
-	// the middleware in front of a protected resource server. The in-process
-	// key provider avoids self-referential HTTP JWKS fetches.
-	validator, err := auth.NewTokenValidator(context.Background(), auth.TokenValidatorConfig{
-		Issuer:   testIssuer,
-		Audience: testAudience,
-	}, auth.WithKeyProvider(&testKeyProvider{key: ts.PrivateKey}))
-	require.NoError(t, err)
-
-	// Chain: audit (outer) -> auth -> bare stub handler. The stub does NOT
-	// re-publish the identity: this test pins that TokenValidator.Middleware
-	// alone publishes the validated identity to the audit holder.
-	auditLog := filepath.Join(t.TempDir(), "audit.log")
-	auditor, err := audit.NewAuditorWithTransport(&audit.Config{LogFile: auditLog}, "streamable-http")
-	require.NoError(t, err)
-	defer auditor.Close()
-
-	stub := http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
-		w.WriteHeader(http.StatusOK)
-	})
-
-	req := httptest.NewRequest(http.MethodPost, "/messages", strings.NewReader("{}"))
-	req.Header.Set("Authorization", "Bearer "+delegated)
-	rr := httptest.NewRecorder()
-	auditor.Middleware(validator.Middleware(stub)).ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code,
-		"delegated token must validate through the auth middleware")
-
-	// Exactly one audit event, carrying the delegation chain.
-	event := readSingleAuditEvent(t, auditLog)
-
-	subjects, ok := event["subjects"].(map[string]any)
-	require.True(t, ok, "subjects should be a map")
-	assert.Equal(t, delegatedUserSub, subjects["user_id"],
-		"audit subject must be the delegated user, not the agent")
-
-	chain, ok := event["delegation_chain"].(map[string]any)
-	require.True(t, ok, "delegation_chain must be present in the audit event")
-	assert.Equal(t, false, chain["truncated"])
-	actors, ok := chain["actors"].([]any)
-	require.True(t, ok, "actors should be an array")
-	require.Len(t, actors, 1, "a single exchange yields a single actor")
-	actor, ok := actors[0].(map[string]any)
-	require.True(t, ok)
-	assert.Equal(t, agentClientID, actor["sub"],
-		"the delegation chain must record the acting agent client")
-}
-
-// TestIntegration_AuditMiddleware_NonDelegatedTokenOmitsDelegationChain is the
-// negative companion to TestIntegration_TokenExchange_AuditDelegationChain: a
-// plain (non-delegated) token through the same audit(auth(stub)) chain must
-// produce an audit event with NO delegation_chain member.
-func TestIntegration_AuditMiddleware_NonDelegatedTokenOmitsDelegationChain(t *testing.T) {
-	t.Parallel()
-
-	const plainUserSub = "audit-plain-user-sub"
-
-	m := startMockOIDC(t)
-	ts := setupTestServerWithMockOIDC(t, m)
-
-	// Mint a plain subject token signed by the server's key (no act claim).
-	signer, err := jose.NewSigner(
-		jose.SigningKey{Algorithm: jose.RS256, Key: ts.PrivateKey},
-		(&jose.SignerOptions{}).WithType("JWT").WithHeader("kid", "test-key"),
-	)
-	require.NoError(t, err)
-
-	now := time.Now()
-	plainToken, err := jwt.Signed(signer).
-		Claims(jwt.Claims{
-			Issuer:   testIssuer,
-			Subject:  plainUserSub,
-			Audience: jwt.Audience{testAudience},
-			Expiry:   jwt.NewNumericDate(now.Add(30 * time.Minute)),
-			IssuedAt: jwt.NewNumericDate(now),
-		}).
-		Serialize()
-	require.NoError(t, err)
-
-	validator, err := auth.NewTokenValidator(context.Background(), auth.TokenValidatorConfig{
-		Issuer:   testIssuer,
-		Audience: testAudience,
-	}, auth.WithKeyProvider(&testKeyProvider{key: ts.PrivateKey}))
-	require.NoError(t, err)
-
-	auditLog := filepath.Join(t.TempDir(), "audit.log")
-	auditor, err := audit.NewAuditorWithTransport(&audit.Config{LogFile: auditLog}, "streamable-http")
-	require.NoError(t, err)
-	defer auditor.Close()
-
-	stub := http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
-		w.WriteHeader(http.StatusOK)
-	})
-
-	req := httptest.NewRequest(http.MethodPost, "/messages", strings.NewReader("{}"))
-	req.Header.Set("Authorization", "Bearer "+plainToken)
-	rr := httptest.NewRecorder()
-	auditor.Middleware(validator.Middleware(stub)).ServeHTTP(rr, req)
-
-	require.Equal(t, http.StatusOK, rr.Code,
-		"plain token must validate through the auth middleware")
-
-	event := readSingleAuditEvent(t, auditLog)
-
-	subjects, ok := event["subjects"].(map[string]any)
-	require.True(t, ok, "subjects should be a map")
-	assert.Equal(t, plainUserSub, subjects["user_id"],
-		"audit subject must be the authenticated user")
-
-	_, exists := event["delegation_chain"]
-	assert.False(t, exists,
-		"a non-delegated token must not produce a delegation_chain member")
 }
 
 // ============================================================================
